Disable All State Edicts
From Excelbel: "It is tedious to turn off your State Edicts one by one. This decision exists to help you with that process. When enacted, all State Edicts in your country will be disabled on the next monthly tick. After that, you'll have to enable this function again to be able to assign State Edicts."